+++ src/transforms/stack.ts
@@ -1,4 +1,5 @@
 import { ok } from "assert";
+import { stringify } from "querystring";
 import { reservedIdentifiers } from "../constants";
 import { ObfuscateOrder } from "../order";
 import { ComputeProbabilityMap } from "../probability";
@@ -15,12 +16,18 @@ import {
   Location,
   MemberExpression,
   Node,
+  RestElement,
+  SequenceExpression,
   SpreadElement,
   VariableDeclaration,
   VariableDeclarator,
   WhileStatement,
 } from "../util/gen";
-import { getIdentifierInfo } from "../util/identifiers";
+import {
+  getDefiningIdentifier,
+  getIdentifierInfo,
+  validateChain,
+} from "../util/identifiers";
 import {
   clone,
   getVarContext,
@@ -30,11 +37,16 @@ import {
   isInBranch,
   prepend,
 } from "../util/insert";
+import { getRandomInteger } from "../util/random";
 import Transform from "./transform";
 
 export default class Stack extends Transform {
+  made: number;
+
   constructor(o) {
     super(o, ObfuscateOrder.Stack);
+
+    this.made = 0;
   }
 
   match(object: Node, parents: Node[]) {
@@ -47,17 +59,15 @@ export default class Stack extends Transform {
 
   transform(object: Node, parents: Node[]) {
     return () => {
-      // SyntaxError: Illegal 'use strict' directive in function with non-simple parameter list
-
-      for (var stmt of object.body.body) {
-        if (stmt.type == "ExpressionStatement") {
-          if (
-            stmt.expression.type == "Literal" &&
-            stmt.expression.value === "use strict"
-          ) {
-            return;
-          }
+      var possible = true;
+      walk(object, parents, (o, p) => {
+        if (o.type == "VariableDeclaration" && o.declarations.length > 1) {
+          possible = false;
+          return "EXIT";
         }
+      });
+      if (!possible) {
+        return;
       }
 
       if (
@@ -69,316 +79,295 @@ export default class Stack extends Transform {
       ) {
         return;
       }
-
       var identifiers: {
-        type: "defined" | "modified" | "reference";
         location: Location;
         value?: Node;
         replacing?: Node;
+        isDefiningNode?: boolean;
       }[] = [];
+      var definingNodes = new Map<Node, Node>();
+      var lastReference = new Map<Node, Node>();
 
-      var firstReference: { [name: string]: Node } = Object.create(null);
-      var lastReference: { [name: string]: Node } = Object.create(null);
-
-      var paramNames = object.params.map((x) => x.name);
-
-      var defined = new Set<string>(paramNames);
-      var references = new Set<string>();
+      var branchMap = new Map<Node, boolean>();
       var illegal = new Set<string>();
+      var queuedReplacements = [];
 
-      var queuedReplaces = new Map<string, [Node, Node][]>();
-
-      var scan = (varNode: Node, varParents: Node[]) => {
-        if (
-          varNode.type == "Identifier" &&
-          !reservedIdentifiers.has(varNode.name) &&
-          !this.options.globalVariables.has(varNode.name)
-        ) {
-          function queue(node1: Node, node2: Node) {
-            if (!queuedReplaces.has(varNode.name)) {
-              queuedReplaces.set(varNode.name, [[node1, node2]]);
-            } else {
-              queuedReplaces.get(varNode.name).push([node1, node2]);
-            }
-          }
-
-          var info = getIdentifierInfo(varNode, varParents);
-
-          if (info.isFunctionParameter || info.isClauseParameter) {
-            illegal.add(varNode.name);
-          } else if (
-            info.spec.isReferenced ||
-            info.spec.isDefined ||
-            info.spec.isModified
-          ) {
-            var value;
-            var replacing = varNode;
-
-            if (info.spec.isDefined) {
-              if (
-                varParents[0].type == "VariableDeclarator" &&
-                varParents[0].id === varNode
-              ) {
-                if (
-                  varParents[2] &&
-                  varParents[2].type == "VariableDeclaration"
-                ) {
-                  if (isForInitialize(varParents[2], varParents.slice(3))) {
-                    illegal.add(varNode.name);
-                    // queue(varParents[2], replacing);
-                  } else if (varParents[2].declarations.length === 1) {
-                    replacing = Identifier("undefined");
-                    value = Identifier("undefined");
-                    if (!varParents[0].init) {
-                      varParents[0].init = Identifier("undefined");
-                    }
-                    replacing = varParents[0].init;
-                    value = { ...varParents[0].init };
-                    queue(varParents[2], ExpressionStatement(replacing));
-                  } else if (varParents[2].declarations.length > 1) {
-                    illegal.add(varNode.name);
-                  }
-                }
-              } else if (
-                varParents[0].type == "FunctionDeclaration" &&
-                varParents[0].id === varNode
-              ) {
-                value = {
-                  ...varParents[0],
-                  type: "FunctionExpression",
-                  id: null,
-                  expression: false,
-                };
-
-                var emptyNode = Identifier("undefined");
-                queue(varParents[0], ExpressionStatement(emptyNode));
-                replacing = emptyNode;
-              } else if (
-                varParents[0].type == "ClassDeclaration" &&
-                varParents[0].id === varNode
-              ) {
-                // console.log(varNode.name, varParents[0].type, info);
-
-                value = {
-                  ...varParents[0],
-                  type: "ClassExpression",
-                  expression: null,
-                };
-
-                var emptyNode = Identifier("undefined");
-                queue(varParents[0], ExpressionStatement(emptyNode));
-                replacing = emptyNode;
-              } else {
-                illegal.add(varNode.name);
-              }
-            }
-            const type =
-              info.spec.isDefined && value
-                ? "defined"
-                : info.spec.isModified
-                ? "modified"
-                : "reference";
-
-            identifiers.push({
-              location: [varNode, varParents],
-              type: type,
-              value: value,
-              replacing: replacing,
-            });
-            lastReference[varNode.name] = varNode;
-
-            // console.log(o.name, info.spec);
-            var isParam = paramNames.includes(varNode.name);
-            if (!firstReference[varNode.name]) {
-              if (!info.spec.isDefined && !isParam) {
-                illegal.add(varNode.name);
-              }
-              firstReference[varNode.name] = varNode;
-            } else if (info.spec.isDefined && !isParam) {
-              illegal.add(varNode.name);
-            }
-
-            if (info.spec.isDefined) {
-              defined.add(varNode.name);
-            } else {
-              references.add(varNode.name);
-            }
-
-            if (
-              info.spec.isModified &&
-              parents.find((x) => x.type == "VariableDeclarator")
-            ) {
-              illegal.add(varNode.name);
-            }
-          }
+      const checkBranch = (o, p, definingNode) => {
+        var isBranch = branchMap.get(definingNode) || isInBranch(o, p, object);
+        if (isBranch) {
+          branchMap.set(definingNode, true);
         }
       };
 
-      walk(object.body, [object, ...parents], (varNode, varParents) => {
-        return () => {
-          scan(varNode, varParents);
-        };
-      });
+      walk(object, parents, (o, p) => {
+        if (o.type == "VariableDeclaration") {
+          return () => {
+            var elements = [];
+            var names = new Set<string>();
+            var forInit = isForInitialize(o, p);
 
-      references.forEach((ref) => {
-        if (!defined.has(ref) || illegal.has(ref)) {
-          identifiers = identifiers.filter((x) => x.location[0].name !== ref);
+            o.declarations.forEach((x) => {
+              if (x.id.type == "Identifier") {
+                var replacing = Literal(0);
+                elements.push(replacing);
+
+                var varParents = [x, o.declarations, o, ...p];
+                identifiers.push({
+                  location: [x.id, varParents],
+                  value: x.init || (!forInit && Identifier("undefined")),
+                  replacing: replacing,
+                });
+
+                definingNodes.set(x.id, x.id);
+                names.add(x.id.name);
+                checkBranch(x.id, varParents, x.id);
+              }
+            });
+
+            var expr: Node = SequenceExpression(elements);
+            if (!forInit) {
+              expr = ExpressionStatement(expr);
+            }
+
+            queuedReplacements.push([o, expr, names]);
+          };
+        }
+        if (o !== object && o.type == "FunctionDeclaration") {
+          return () => {
+            var value = {
+              ...o,
+              type: "FunctionExpression",
+              expression: false,
+              id: null,
+            };
+            var replacing = Literal(0);
+
+            identifiers.push({
+              location: [o.id, [o, ...p]],
+              replacing: replacing,
+              value: value,
+            });
+            definingNodes.set(o.id, o.id);
+            checkBranch(o.id, [o, ...p], o.id);
+
+            queuedReplacements.push([
+              o,
+              ExpressionStatement(replacing),
+              new Set([o.id.name]),
+            ]);
+          };
+        }
+        if (o !== object && o.type == "ClassDeclaration") {
+          return () => {
+            var value = {
+              ...o,
+              type: "ClassDeclaration",
+            };
+            var replacing = Literal(0);
+
+            identifiers.push({
+              location: [o.id, [o, ...p]],
+              replacing: replacing,
+              value: value,
+            });
+            definingNodes.set(o.id, o.id);
+            checkBranch(o.id, [o, ...p], o.id);
+
+            queuedReplacements.push([
+              o,
+              ExpressionStatement(replacing),
+              new Set([o.id.name]),
+            ]);
+          };
+        }
+
+        if (
+          o !== object.id &&
+          o.type == "Identifier" &&
+          !reservedIdentifiers.has(o.name) &&
+          !this.options.globalVariables.has(o.name)
+        ) {
+          var sliced = p.slice(0, p.indexOf(object));
+          var varIndex = sliced.findIndex(
+            (x) => x.type == "VariableDeclarator"
+          );
+          if (
+            varIndex !== -1 &&
+            sliced[varIndex].id == (sliced[varIndex - 1] || o)
+          ) {
+            return;
+          }
+
+          var info = getIdentifierInfo(o, p);
+          if (info.spec.isDefined || info.spec.isReferenced) {
+            if (info.isFunctionDeclaration || info.isClassDeclaration) {
+              return;
+            }
+            if (info.isFunctionParameter) {
+              var fn = p.find((x) => isFunction(x));
+              if (fn == object) {
+              } else {
+                illegal.add(o.name);
+              }
+              return;
+            }
+            if (info.isClauseParameter) {
+              illegal.add(o.name);
+              return;
+            }
+
+            var definingNode = info.spec.isDefined ? o : null;
+
+            if (!definingNode) {
+              var location = getDefiningIdentifier(o, p);
+              if (location && location[1].includes(object)) {
+                definingNode = location[0];
+              }
+            }
+
+            if (definingNode) {
+              definingNodes.set(o, definingNode);
+
+              lastReference.set(definingNode, o);
+              identifiers.push({ location: [o, p] });
+            }
+          }
         }
       });
-      illegal.forEach((illegal) => {
-        queuedReplaces.delete(illegal);
-        identifiers = identifiers.filter((x) => x.location[0].name !== illegal);
-      });
 
-      // console.log(object.id.name, identifiers.length);
+      illegal.forEach((bad) => {
+        identifiers = identifiers.filter((x) => x.location[0].name !== bad);
+      });
 
       if (!identifiers.length) {
         return;
       }
 
-      // console.log(defined);
-
-      var rollback = clone(object);
-      queuedReplaces.forEach((value) => {
-        value.forEach(([node1, node2]) => {
-          this.replace(node1, node2);
+      queuedReplacements.forEach((x) => {
+        var can = true;
+        x[2].forEach((x) => {
+          if (illegal.has(x)) {
+            // console.log(x);
+            can = false;
+          }
         });
+        if (can) {
+          this.replace(x[0], x[1]);
+        }
       });
 
+      // console.log(illegal);
+
+      var mappings: string[] = [...object.params.map((x) => x.name)];
+
+      // Array(getRandomInteger(1, 4))
+      //   .fill(0)
+      //   .forEach((x) => {
+      //     mappings.push(this.getPlaceholder());
+      //   });
+
+      var startingSize = mappings.length;
       var stackName = this.getPlaceholder();
 
-      // stack is default all params in an array
-      var mappings: string[] = [...paramNames];
-
-      for (var i = 0; i < identifiers.length; i++) {
-        var { type, location, value, replacing } = identifiers[i];
+      for (var identifier of identifiers) {
+        var { location, value, replacing, isDefiningNode } = identifier;
         var [varNode, varParents] = location;
 
-        if (typeof varNode.name !== "string") {
-          this.replace(object, rollback);
-          return;
+        if (!replacing) {
+          replacing = varNode;
         }
 
-        var isFirstTimeUsed = firstReference[varNode.name] === varNode;
-        var isLastTimeUsed = lastReference[varNode.name] === varNode;
-
-        var isParam = paramNames.includes(varNode.name);
-
-        var index = mappings.findIndex((x) => x === varNode.name);
-        if (index === -1) {
-          index = undefined;
-
-          if (isFirstTimeUsed) {
-            index = mappings.length;
-            mappings.push(varNode.name);
-          } else {
-            this.replace(object, rollback);
-            return;
-          }
+        if (illegal.has(varNode.name)) {
+          continue;
         }
 
-        var isFirstIndex =
-          mappings.length === 0 || mappings[0] === varNode.name;
-        var isLastIndex =
-          mappings.length === 0 ||
-          mappings[mappings.length - 1] === varNode.name;
+        var index = mappings.indexOf(varNode.name);
 
-        var isBranch = isInBranch(varNode, varParents, object);
+        var definingNode = definingNodes.get(varNode);
+        var exitingNode = lastReference.get(definingNode);
 
+        var isFirstTimeUsed = isDefiningNode || definingNode == varNode;
+        var isLastTimeUsed = exitingNode == varNode;
+
+        var isBranch = branchMap.get(definingNode);
+
+        if (index == -1) {
+          index = mappings.length;
+          mappings.push(varNode.name);
+        }
+
+        var isFirstIndex = index === 0;
+        var isLastIndex = index === mappings.length - 1;
+
+        // console.log(varNode, definingNode, exitingNode);
         // console.log({
         //   name: varNode.name,
-        //   index,
-        //   isFirstTimeUsed,
-        //   isLastTimeUsed,
         //   isFirstIndex,
         //   isLastIndex,
-        //   isBranch,
-        //   isParam,
+        //   isFirstTimeUsed,
+        //   isLastTimeUsed,
         // });
 
-        if (isFirstTimeUsed && !isParam) {
-          // add to the array
+        var lastShortcut =
+          isLastTimeUsed &&
+          !isBranch &&
+          (isFirstIndex ? "shift" : isLastIndex ? "pop" : null);
+        var firstShortcut =
+          isFirstTimeUsed &&
+          !isBranch &&
+          (isFirstIndex ? "unshift" : isLastIndex ? "push" : null);
 
-          if (type == "defined") {
-            ok(value);
-            ok(replacing);
+        if (
+          varParents[0].type == "AssignmentExpression" &&
+          varParents[0].operator == "=" &&
+          varParents[0].left === varNode
+        ) {
+          value = varParents[0].right;
+          replacing = varParents[0];
+        }
 
-            if ((isFirstIndex || isLastIndex) && !isBranch) {
-              this.replace(
-                replacing,
-                CallExpression(
-                  MemberExpression(
-                    Identifier(stackName),
-                    Identifier(isFirstIndex ? "unshift" : "push"),
-                    false
-                  ),
-                  [value]
-                )
-              );
-              continue;
-            }
-
+        if (value && firstShortcut) {
+          this.replace(
+            replacing,
+            CallExpression(
+              MemberExpression(
+                Identifier(stackName),
+                Identifier(firstShortcut),
+                false
+              ),
+              [clone(value)]
+            )
+          );
+        } else if (lastShortcut) {
+          mappings = mappings.filter((x) => x != varNode.name);
+          this.replace(
+            replacing,
+            Template(`${stackName}.${lastShortcut}()`).single().expression
+          );
+        } else if (typeof index === "number") {
+          if (value) {
             this.replace(
               replacing,
               AssignmentExpression(
                 "=",
                 MemberExpression(Identifier(stackName), Literal(index), true),
-                value
+                clone(value) || Identifier("undefined")
               )
             );
-            continue;
-          }
-        }
-
-        if (isLastTimeUsed && typeof index === "number") {
-          if ((isFirstIndex || isLastIndex) && !isBranch) {
-            // remove from array
-
-            mappings = mappings.filter((x) => x !== varNode.name);
-
-            // console.log(varNode.name, "->", isFirstIndex ? "shift" : "pop");
+          } else {
             this.replace(
               replacing,
-              CallExpression(
-                MemberExpression(
-                  Identifier(stackName),
-                  Identifier(isFirstIndex ? "shift" : "pop"),
-                  false
-                ),
-                []
-              )
+              MemberExpression(Identifier(stackName), Literal(index), true)
             );
-            continue;
           }
-        }
-
-        if (typeof index === "number") {
-          this.replace(
-            replacing,
-            MemberExpression(Identifier(stackName), Literal(index), true)
-          );
-        } else {
-          this.replace(object, rollback);
-          return;
         }
       }
 
-      // if (paramNames.length) {
-      object.params = [SpreadElement(Identifier(stackName))];
+      object.params = [RestElement(Identifier(stackName))];
       prepend(
         object.body,
-        Template(`
-        ${stackName}.length = ${paramNames.length};
-      `).single()
+        Template(`${stackName}.length = ${startingSize}`).single()
       );
-      // } else {
-      // prepend(
-      //   object.body,
-      //   VariableDeclaration(
-      //     VariableDeclarator(Identifier(stackName), ArrayExpression([]))
-      //   )
-      // );
-      // }
     };
   }
 }