+++ test/code/Cash.test.ts
@@ -2,26 +2,27 @@ import { readFileSync, writeFileSync } from "fs";
 import { join } from "path";
 import JsConfuser from "../../src/index";
 
-var CASH_JS = readFileSync(join(__dirname, "./Cash.src"), "utf-8");
+var CASH_JS = readFileSync(join(__dirname, "./Cash.src.js"), "utf-8");
 
 it("works with Cash.js on High Preset", async () => {
   var output = await JsConfuser(CASH_JS, {
     target: "browser",
     preset: "high",
-    // stringEncoding: true,
-    // stringConcealing: true,
-    // stringSplitting: true,
-    // controlFlowFlattening: true,
-    // deadCode: true,
-    // duplicateLiteralsRemoval: true,
-    // calculator: true,
-    // shuffle: "hash",
-    // objectExtraction: true,
-    // renameVariables: true,
+    stringEncoding: false,
+    stringConcealing: false,
+    stringSplitting: false,
+    controlFlowFlattening: false,
+    deadCode: false,
+    duplicateLiteralsRemoval: false,
+    calculator: false,
+    shuffle: false,
+    objectExtraction: false,
+    renameVariables: false,
     // stack: true,
-    // dispatcher: true,
-    // flatten: true,
-    // controlFlowFlattening: true,
+    dispatcher: false,
+    flatten: false,
+    compact: false,
+    globalConcealing: false,
   });
 
   // Make the required document variables for initialization
@@ -47,7 +48,9 @@ it("works with Cash.js on High Preset", async () => {
   } as any;
   window.window = window;
 
-  // writeFileSync(join(__dirname, "Cash.output"), output, { encoding: "utf-8" });
+  writeFileSync(join(__dirname, "Cash.output.js"), output, {
+    encoding: "utf-8",
+  });
 
   eval(output);